Added a file that tries to extract pure noise sounds from non pure noise sounds.
Once we can automate the process of gathering a sample of pure noise from a soundfile, we can derive the noise signature and that will help us intensely to denoise the audiofile. My best guess up to now is to use a spectral noise gate to do the denoising, once we have a noise signature.